backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Email Processing
imapclient>=3.0.0
email-validator>=2.2.0

# PDF Processing
PyPDF2>=3.0.1
pdf2image>=1.17.0
Pillow>=10.4.0

# OCR
pytesseract>=0.3.13

# GenAI
openai>=1.51.0
google-generativeai>=0.8.0

# Auth
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Utilities
aiofiles>=24.1.0
httpx>=0.27.0
numpy>=1.26.0
